Use NO_COLOR and smarter "if it's GNU" test
See https://no-color.org. Indeed some terminals like `xterm-mono` return
nothing for `terminfo[colors]`.

To test "if it's GNU", don't just test for the presence of `dircolors`,
but also check if `ls` is GNU by accepting the `--version` parameter.

This is being done because MacOS's homebrew allowed installing a mixture
of GNU and BSD tools together, where we got the GNU `dircolors`, but the
BSD `ls`. See Homebrew/homebrew-core#36241

The homebrew changes were partially reverted in
Homebrew/homebrew-core#36494, which does not mix the two tools above
anymore, thus not breaking our script anymore either. But I prefer we
keep this safer for the future.

Current changes didn't show any change in the benchmarks of the default
Zim installation.

Closes zimfw/zimfw#328.